Add files via upload
sort of almost unfeasably janky version of client code that uses infinite threads (until the program crashes at some point), must stall between each thread (or the program will crash), and far to many global variables, but it does send and receive messages in a timely manor and lets one client send as many messages as they want before getting a reply (that is until it crashes of course). you also need to enter EXIT twice for it to work for some reason.