server using non blocking sockets and with signal handler (not perfect)
The issue appears to be that whenever Ctrl C is pressed on the server terminal it makes the select function that something has happened on both the master socket and the client socket, so once it expects something it gets stuck since that something doesn't exist (it was faked out by the signal handler somehow). I tried using a global Boolean to keep track of whether Ctrl C had been pushed, and I tried making sd a non blocking socket (although I suspect the client code means it already is anyway) and none of that worked.