implemented & for multiple commands on one line
I think this is all the things it needs to do. My solution for the & operator is pretty scuffed tho since you would need to do a bit refactoring to implement something like a pipe operator (since you'd need a function that could tokenize with more than one deliminator at once, and then do different things with different token groups based on their deliminator) but I guess all we need is the & so it should be okay.